requirements.txt
aiohttp
al-utils-almirai
